update chatapi function, add docs, move types
