+++ apps/src/aichat/types.ts
@@ -1,4 +1,3 @@
-import {Role, Status} from './constants';
 // TODO: Ideally this type would only contain keys present in
 // translated string JSON files (ex. apps/i18n/aichat/en_us.json).
 // However, this requires depending on files outside of apps/src,
@@ -16,3 +15,16 @@ export type ChatCompletionMessage = {
   chatMessageText: string;
   status: Status;
 };
+
+export enum Role {
+  ASSISTANT = 'assistant',
+  USER = 'user',
+  SYSTEM = 'system',
+}
+
+export enum Status {
+  OK = 'ok',
+  PERSONAL = 'personal',
+  INAPPROPRIATE = 'inappropriate',
+  UNKNOWN = 'unknown',
+}
